Separate StationStatus.vue
